+++ tests/nativeapp/test_manager.py
@@ -51,6 +51,7 @@ from snowflake.cli._plugins.stage.diff import (
     DiffResult,
     StagePathType,
 )
+from snowflake.cli._plugins.stage.manager import DefaultStagePathParts
 from snowflake.cli._plugins.workspace.manager import WorkspaceManager
 from snowflake.cli.api.console import cli_console as cc
 from snowflake.cli.api.entities.common import EntityActions
@@ -146,12 +147,11 @@ def test_sync_deploy_root_with_stage(
         console=cc,
         deploy_root=dm.project_root / pkg_model.deploy_root,
         package_name=package_name,
-        stage_schema=stage_schema,
         bundle_map=mock_bundle_map,
         role="new_role",
         prune=True,
         recursive=True,
-        stage_fqn=stage_fqn,
+        stage_path=DefaultStagePathParts.from_fqn(stage_fqn),
     )
 
     mock_stage_exists.assert_called_once_with(stage_fqn)
@@ -159,13 +159,78 @@ def test_sync_deploy_root_with_stage(
         mock_create_schema.assert_called_once_with(stage_schema, database=package_name)
         mock_create_stage.assert_called_once_with(stage_fqn)
     mock_compute_stage_diff.assert_called_once_with(
-        dm.project_root / pkg_model.deploy_root, "app_pkg.app_src.stage"
+        local_root=dm.project_root / pkg_model.deploy_root,
+        stage_path=DefaultStagePathParts.from_fqn("app_pkg.app_src.stage"),
     )
     mock_local_diff_with_stage.assert_called_once_with(
         role="new_role",
         deploy_root_path=dm.project_root / pkg_model.deploy_root,
         diff_result=mock_diff_result,
-        stage_fqn="app_pkg.app_src.stage",
+        stage_full_path="app_pkg.app_src.stage",
+    )
+
+
+@mock.patch(SQL_FACADE_STAGE_EXISTS)
+@mock.patch(SQL_FACADE_CREATE_SCHEMA)
+@mock.patch(SQL_FACADE_CREATE_STAGE)
+@mock.patch(f"{ENTITIES_UTILS_MODULE}.compute_stage_diff")
+@mock.patch(f"{ENTITIES_UTILS_MODULE}.sync_local_diff_with_stage")
+@pytest.mark.parametrize("stage_exists", [True, False])
+def test_sync_deploy_root_with_stage_subdir(
+    mock_local_diff_with_stage,
+    mock_compute_stage_diff,
+    mock_create_stage,
+    mock_create_schema,
+    mock_stage_exists,
+    temp_dir,
+    mock_cursor,
+    stage_exists,
+):
+    mock_stage_exists.return_value = stage_exists
+    mock_diff_result = DiffResult(different=[StagePathType("setup.sql")])
+    mock_compute_stage_diff.return_value = mock_diff_result
+    mock_local_diff_with_stage.return_value = None
+    current_working_directory = os.getcwd()
+    create_named_file(
+        file_name="snowflake.yml",
+        dir_name=current_working_directory,
+        contents=[mock_snowflake_yml_file_v2],
+    )
+    dm = _get_dm()
+    # add subdir replace
+    dm.project_definition.entities["app_pkg"].stage_subdirectory = "v1"
+    pkg_model: ApplicationPackageEntityModel = dm.project_definition.entities["app_pkg"]
+
+    assert mock_diff_result.has_changes()
+    mock_bundle_map = mock.Mock(spec=BundleMap)
+    package_name = pkg_model.fqn.name
+    stage_fqn = f"{package_name}.{pkg_model.stage}"
+    stage_full_path = f"{stage_fqn}/v1"
+    stage_schema = extract_schema(stage_fqn)
+    sync_deploy_root_with_stage(
+        console=cc,
+        deploy_root=dm.project_root / pkg_model.deploy_root,
+        package_name=package_name,
+        bundle_map=mock_bundle_map,
+        role="new_role",
+        prune=True,
+        recursive=True,
+        stage_path=DefaultStagePathParts.from_fqn(stage_fqn, "v1"),
+    )
+
+    mock_stage_exists.assert_called_once_with(stage_fqn)
+    if not stage_exists:
+        mock_create_schema.assert_called_once_with(stage_schema, database=package_name)
+        mock_create_stage.assert_called_once_with(stage_fqn)
+    mock_compute_stage_diff.assert_called_once_with(
+        local_root=dm.project_root / pkg_model.deploy_root,
+        stage_path=DefaultStagePathParts.from_fqn(stage_fqn, "v1"),
+    )
+    mock_local_diff_with_stage.assert_called_once_with(
+        role="new_role",
+        deploy_root_path=dm.project_root / pkg_model.deploy_root,
+        diff_result=mock_diff_result,
+        stage_full_path=stage_full_path,
     )
 
 
@@ -216,12 +281,11 @@ def test_sync_deploy_root_with_stage_prune(
         console=mock_console,
         deploy_root=dm.project_root / pkg_model.deploy_root,
         package_name=package_name,
-        stage_schema=extract_schema(stage_fqn),
         bundle_map=mock_bundle_map,
         role="new_role",
         prune=prune,
         recursive=True,
-        stage_fqn=stage_fqn,
+        stage_path=DefaultStagePathParts.from_fqn(stage_fqn),
     )
 
     if expected_warn:
@@ -1383,7 +1447,9 @@ def test_validate_use_scratch_stage(mock_execute, mock_deploy, temp_dir, mock_cu
         paths=[],
         print_diff=False,
         validate=False,
-        stage_fqn=f"{pkg_model.fqn.name}.{pkg_model.scratch_stage}",
+        stage_path=DefaultStagePathParts.from_fqn(
+            f"{pkg_model.fqn.name}.{pkg_model.scratch_stage}"
+        ),
         interactive=False,
         force=True,
         run_post_deploy_hooks=False,
@@ -1460,7 +1526,9 @@ def test_validate_failing_drops_scratch_stage(
         paths=[],
         print_diff=False,
         validate=False,
-        stage_fqn=f"{pkg_model.fqn.name}.{pkg_model.scratch_stage}",
+        stage_path=DefaultStagePathParts.from_fqn(
+            f"{pkg_model.fqn.name}.{pkg_model.scratch_stage}"
+        ),
         interactive=False,
         force=True,
         run_post_deploy_hooks=False,