+++ tests/nativeapp/test_sf_sql_facade.py
@@ -1799,7 +1799,7 @@ def test_upgrade_application_unversioned(
         sql_facade.upgrade_application(
             name=app_name,
             install_method=SameAccountInstallMethod.unversioned_dev(),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=None,
             should_authorize_event_sharing=None,
             role=role,
@@ -1850,7 +1850,7 @@ def test_upgrade_application_version_and_patch(
         sql_facade.upgrade_application(
             name=app_name,
             install_method=SameAccountInstallMethod.versioned_dev("3", 2),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=True,
             should_authorize_event_sharing=True,
             role=role,
@@ -1897,7 +1897,7 @@ def test_upgrade_application_from_release_directive(
         sql_facade.upgrade_application(
             name=app_name,
             install_method=SameAccountInstallMethod.release_directive(),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=True,
             should_authorize_event_sharing=True,
             role=role,
@@ -1943,7 +1943,7 @@ def test_upgrade_application_converts_expected_programmingerrors_to_user_errors(
         sql_facade.upgrade_application(
             name=app_name,
             install_method=SameAccountInstallMethod.unversioned_dev(),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=True,
             should_authorize_event_sharing=True,
             role=role,
@@ -2008,7 +2008,7 @@ def test_upgrade_application_special_message_for_event_sharing_error(
         sql_facade.upgrade_application(
             name=app_name,
             install_method=SameAccountInstallMethod.versioned_dev("v1"),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=False,
             should_authorize_event_sharing=False,
             role=role,
@@ -2057,7 +2057,7 @@ def test_upgrade_application_converts_unexpected_programmingerrors_to_unclassifi
         sql_facade.upgrade_application(
             name=app_name,
             install_method=SameAccountInstallMethod.unversioned_dev(),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=True,
             should_authorize_event_sharing=True,
             role=role,
@@ -2106,7 +2106,7 @@ def test_upgrade_application_with_release_channel_same_as_app_properties(
         sql_facade.upgrade_application(
             name=app_name,
             install_method=SameAccountInstallMethod.release_directive(),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=True,
             should_authorize_event_sharing=True,
             role=role,
@@ -2138,7 +2138,7 @@ def test_upgrade_application_with_release_channel_not_same_as_app_properties_the
         sql_facade.upgrade_application(
             name=app_name,
             install_method=SameAccountInstallMethod.release_directive(),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=True,
             should_authorize_event_sharing=True,
             role=role,
@@ -2193,7 +2193,7 @@ def test_create_application_with_minimal_clauses(
             name=app_name,
             package_name=pkg_name,
             install_method=SameAccountInstallMethod.release_directive(),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=None,
             should_authorize_event_sharing=None,
             role=role,
@@ -2245,7 +2245,7 @@ def test_create_application_with_all_clauses(
             name=app_name,
             package_name=pkg_name,
             install_method=SameAccountInstallMethod.unversioned_dev(),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=True,
             should_authorize_event_sharing=True,
             role=role,
@@ -2298,7 +2298,7 @@ def test_create_application_converts_expected_programmingerrors_to_user_errors(
             name=app_name,
             package_name=pkg_name,
             install_method=SameAccountInstallMethod.release_directive(),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=None,
             should_authorize_event_sharing=None,
             role=role,
@@ -2360,7 +2360,7 @@ def test_create_application_special_message_for_event_sharing_error(
             name=app_name,
             package_name=pkg_name,
             install_method=SameAccountInstallMethod.versioned_dev("3", 1),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=False,
             should_authorize_event_sharing=False,
             role=role,
@@ -2418,7 +2418,7 @@ def test_create_application_converts_unexpected_programmingerrors_to_unclassifie
             name=app_name,
             package_name=pkg_name,
             install_method=SameAccountInstallMethod.release_directive(),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=None,
             should_authorize_event_sharing=None,
             role=role,
@@ -2471,7 +2471,7 @@ def test_create_application_with_release_channel(
             name=app_name,
             package_name=pkg_name,
             install_method=SameAccountInstallMethod.release_directive(),
-            stage_fqn=stage_fqn,
+            path_to_version_directory=stage_fqn,
             debug_mode=None,
             should_authorize_event_sharing=None,
             role=role,
@@ -3537,7 +3537,7 @@ def test_create_version_in_package(
                 package_name=package_name,
                 version=version,
                 role=role,
-                stage_fqn=stage_fqn,
+                path_to_version_directory=stage_fqn,
             )
 
 
@@ -3577,7 +3577,7 @@ def test_create_version_in_package_with_label(
                 package_name=package_name,
                 version=version,
                 role=role,
-                stage_fqn=stage_fqn,
+                path_to_version_directory=stage_fqn,
                 label=label,
             )
 
@@ -3616,7 +3616,7 @@ def test_create_version_with_special_characters(
                 package_name=package_name,
                 version=version,
                 role=role,
-                stage_fqn=stage_fqn,
+                path_to_version_directory=stage_fqn,
             )
 
 
@@ -3637,7 +3637,7 @@ def test_create_version_in_package_with_error(
                 package_name=package_name,
                 version=version,
                 role=role,
-                stage_fqn=stage_fqn,
+                path_to_version_directory=stage_fqn,
             )