+++ tests_integration/nativeapp/test_version.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from shlex import split
+from textwrap import dedent
 from typing import Any, Union
 
 import yaml
@@ -29,7 +30,7 @@ from tests.nativeapp.factories import (
 )
 from tests.project.fixtures import *
 from tests_integration.test_utils import contains_row_with, row_from_snowflake_session
-
+from tests_integration.testing_utils.project_fixtures import setup_v2_project_w_subdir
 
 # A minimal set of fields to compare when checking version output
 VERSION_FIELDS_TO_OUTPUT = [
@@ -79,11 +80,24 @@ def temporary_role(request, snowflake_session, resource_suffix):
 # Tests a simple flow of an existing project, executing snow app version create, drop and teardown, all with distribution=internal
 @pytest.mark.integration
 @pytest.mark.parametrize(
-    "create_command,list_command,drop_command,test_project",
+    "project_name,create_command,list_command,drop_command,test_project",
     [
-        ["app version create", "app version list", "app version drop", "napp_init_v1"],
-        ["app version create", "app version list", "app version drop", "napp_init_v2"],
         [
+            "myapp",
+            "app version create",
+            "app version list",
+            "app version drop",
+            "napp_init_v1",
+        ],
+        [
+            "myapp",
+            "app version create",
+            "app version list",
+            "app version drop",
+            "napp_init_v2",
+        ],
+        [
+            "myapp",
             "ws version create --entity-id=pkg",
             "ws version list --entity-id=pkg",
             "ws version drop --entity-id=pkg",
@@ -97,13 +111,58 @@ def test_nativeapp_version_create_and_drop(
     default_username,
     resource_suffix,
     nativeapp_project_directory,
+    project_name,
     create_command,
     list_command,
     drop_command,
     test_project,
 ):
-    project_name = "myapp"
     with nativeapp_project_directory(test_project):
+        result_create = runner.invoke_with_connection_json(
+            [*split(create_command), "v1", "--force", "--skip-git-check"]
+        )
+        assert result_create.exit_code == 0
+
+        # package exist
+        package_name = f"{project_name}_pkg_{default_username}{resource_suffix}".upper()
+        assert contains_row_with(
+            row_from_snowflake_session(
+                snowflake_session.execute_string(
+                    f"show application packages like '{package_name}'",
+                )
+            ),
+            dict(name=package_name),
+        )
+
+        # app package contains version v1
+        expect = snowflake_session.execute_string(
+            f"show versions in application package {package_name}"
+        )
+        actual = runner.invoke_with_connection_json(split(list_command))
+        assert actual.json == row_from_snowflake_session(expect)
+
+        result_drop = runner.invoke_with_connection_json(
+            [*split(drop_command), "v1", "--force"]
+        )
+        assert result_drop.exit_code == 0
+        actual = runner.invoke_with_connection_json(split(list_command))
+        assert len(actual.json) == 0
+
+
+@pytest.mark.integration
+def test_nativeapp_version_create_and_drop_stage_subdir(
+    runner,
+    snowflake_session,
+    default_username,
+    resource_suffix,
+    nativeapp_teardown,
+    setup_v2_project_w_subdir,
+):
+    project_name, _ = setup_v2_project_w_subdir()
+    drop_command = "app version drop --package-entity-id=pkg_v1"
+    list_command = "app version list --package-entity-id=pkg_v1"
+    create_command = "app version create --package-entity-id=pkg_v1"
+    with nativeapp_teardown():
         result_create = runner.invoke_with_connection_json(
             [*split(create_command), "v1", "--force", "--skip-git-check"]
         )
@@ -170,6 +229,49 @@ def test_nativeapp_upgrade(
         assert actual.json == row_from_snowflake_session(expect)
 
         runner.invoke_with_connection_json(["app", "run", "--version", "v1", "--force"])
+
+        expect = row_from_snowflake_session(
+            snowflake_session.execute_string(f"desc application {app_name}")
+        )
+        assert contains_row_with(expect, {"property": "name", "value": app_name})
+        assert contains_row_with(expect, {"property": "version", "value": "V1"})
+        assert contains_row_with(expect, {"property": "patch", "value": "0"})
+
+        runner.invoke_with_connection_json([*split(drop_command), "v1", "--force"])
+
+
+@pytest.mark.integration
+def test_nativeapp_upgrade_w_subdirs(
+    runner,
+    snowflake_session,
+    default_username,
+    resource_suffix,
+    nativeapp_teardown,
+    setup_v2_project_w_subdir,
+):
+    project_name, _ = setup_v2_project_w_subdir()
+    create_command = (
+        "app version create v1 --package-entity-id=pkg_v1 --force --skip-git-check"
+    )
+    list_command = "app version list --package-entity-id=pkg_v1"
+    drop_command = "app version drop --package-entity-id=pkg_v1"
+    with nativeapp_teardown():
+        runner.invoke_with_connection_json(["app", "run", "--app-entity-id=app_v1"])
+        runner.invoke_with_connection_json(split(create_command))
+
+        # package exist
+        package_name = f"{project_name}_pkg_{default_username}{resource_suffix}".upper()
+        app_name = f"{project_name}_app_v1_{default_username}{resource_suffix}".upper()
+        # app package contains version v1
+        expect = snowflake_session.execute_string(
+            f"show versions in application package {package_name}"
+        )
+        actual = runner.invoke_with_connection_json(split(list_command))
+        assert actual.json == row_from_snowflake_session(expect)
+
+        runner.invoke_with_connection_json(
+            ["app", "run", "--app-entity-id=app_v1", "--version", "v1", "--force"]
+        )
 
         expect = row_from_snowflake_session(
             snowflake_session.execute_string(f"desc application {app_name}")