+++ RELEASE-NOTES.md
@@ -28,6 +28,7 @@
 * `snow app version create` now returns version, patch, and label in JSON format.
 * Add ability to specify release channel when creating application instance from release directive: `snow app run --from-release-directive --channel=<channel>`
 * Add ability to list release channels through `snow app release-channel list` command
+* Add an optional `stage_subdirectory` field to `application package` entity. When specified, application artifacts are uploaded to this subdirectory instead of the root of the application package's stage.
 
 ## Fixes and improvements
 * Fixed crashes with older x86_64 Intel CPUs.